refactor(mobile): capa de datos en centavos enteros
Sprint flutter-integer-cents-v1, fase 2 de 5.

**FinancialStateService**: todos los balances derivados pasan a `int`
(`watchAccountBalance`, `accountBalanceNow`, `watchBo/De/Cr`,
`watchTotalLoans`, `accountBalanceAtomic`, `_ReplayBalanceStream`). El
SQL agregado ahora suma INTEGER, así que `SUM()` es exacto por
construcción — el error de precisión de los balances desaparece de raíz.

**DAOs**: signatures monetarias `double → int` en AccountsDao
(creditLimit), CategoriesDao (monthlyLimit), LoansDao (principalAmount,
monthlyPayment, balanceOf, watchBalance), WeeklyBudgetsDao (amount,
watchBudgetBalance) y EntriesDao (amount, principalAmount,
interestAmount, minAmount, maxAmount).

**Se eliminan las 3 tolerancias `+ 0.005` (RF-010, RN-IC-05)**:
- `overpay_debt` en registerDebtPayment — el parche del hotfix
  0.32.1+120 que motivó este sprint.
- `overpay_loan` en registerLoanPayment y updateLoanPayment.
Con centavos enteros la resta es exacta y `>` recupera su semántica
literal. La validación del split pasa de `.abs() >= 0.005` a igualdad
exacta `principal + interest != amount`.

**Se eliminan los guards `isFinite` del hotfix F-SEC-01**: un `int` de
Dart no puede ser NaN ni Infinity, así que la clase de bug que cubrían
("NaN persistido en BD que corrompe balanceOf") es inexpresable. El
guard sobrevive donde sigue teniendo sentido: sobre el `double` crudo
de un payload de backup v1/v2, antes de convertirlo.

**ReportsService**: agregaciones y campos monetarios a `int`. Lo que
sigue siendo `double` con intención: `percent`/`usedPct` (porcentajes),
`p25/p50/p75` (umbrales de cuartil **interpolados** entre dos valores —
no son montos que el usuario vea). Los promedios mensuales
(`historicalAverage`) se redondean al centavo para no reintroducir
fracciones en el dominio.

**EntriesFilters**: `minAmount`/`maxAmount` a `int`. Las vistas guardadas
persisten filtros como JSON y **sobreviven a la migración**, así que
`_tryParseSavedCents` discrimina por tipo JSON: un `double` (`100.0`) es
payload legacy en unidades → `centsFromDouble`; un `int` (`10000`) ya
son centavos. Sin esto, una vista guardada con "mínimo $100" pasaría a
filtrar por $1 tras actualizar.

+++ mobile/lib/data/daos/accounts_dao.dart
@@ -84,7 +84,7 @@ class AccountsDao extends DatabaseAccessor<FincoreDatabase>
     required String name,
     required String type, // 'debit' | 'credit'
     String? description,
-    double? creditLimit,
+    int? creditLimit,
     int? closingDay,
     int? paymentDay,
     double? interestRate,
@@ -176,7 +176,7 @@ class AccountsDao extends DatabaseAccessor<FincoreDatabase>
     required String id,
     String? name,
     String? description,
-    double? creditLimit,
+    int? creditLimit,
     int? closingDay,
     int? paymentDay,
     double? interestRate,
@@ -374,7 +374,7 @@ class AccountsDao extends DatabaseAccessor<FincoreDatabase>
   void _validateCreditMetadata({
     int? closingDay,
     int? paymentDay,
-    double? creditLimit,
+    int? creditLimit,
   }) {
     if (closingDay != null && (closingDay < 1 || closingDay > 31)) {
       throw const AccountsDaoError(
